requirements.txt
azure==1.0.3
azure-common==1.1.4
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==1.0.0
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==1.0.0
azure-servicebus==0.20.1
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
codegen==1.0
Django==1.10.3
django-analytical==2.2.2
django-pipeline==1.6.9
django-pipeline-browserify==0.5.0
django-registration==2.2
django-storages==1.5.1
django-webpack-loader==0.4.1
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
futures==3.0.5
Pillow==3.4.2
psycopg2==2.6.2
PyExecJS==1.4.0
PyJWT==1.4.2
PyReact==0.6.0
python-dateutil==2.6.0
python-http-client==2.2.1
PyYAML==3.12
requests==2.12.4
sendgrid==3.6.3
six==1.10.0
stripe==1.53.0
uWSGI==2.0.14
